Refactored ResourceInfo and ResourceInfoFactory Unit and Integration tests
