+++ tests/Integration/SDK/ResourceInfoTest.php
@@ -1,0 +1,56 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OpenTelemetry\Tests\Integration\SDK\Resource;
+
+use AssertWell\PHPUnitGlobalState\EnvironmentVariables;
+use OpenTelemetry\SDK\Common\Attribute\Attributes;
+use OpenTelemetry\SDK\Resource\Detectors;
+use OpenTelemetry\SDK\Resource\ResourceInfo;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @coversNothing
+ */
+class ResourceInfoTest extends TestCase
+{
+    use EnvironmentVariables;
+
+    public function tearDown(): void
+    {
+        $this->restoreEnvironmentVariables();
+    }
+
+    /**
+     * @dataProvider environmentResourceProvider
+     * @group compliance
+     */
+    public function test_resource_from_environment(string $envAttributes, array $userAttributes, array $expected): void
+    {
+        $this->setEnvironmentVariable('OTEL_RESOURCE_ATTRIBUTES', $envAttributes);
+        $resource = (new Detectors\Composite([
+            new Detectors\Constant(ResourceInfo::create(new Attributes($userAttributes))),
+            new Detectors\Environment(),
+        ]))->getResource();
+        foreach ($expected as $name => $value) {
+            $this->assertSame($value, $resource->getAttributes()->get($name));
+        }
+    }
+
+    public function environmentResourceProvider()
+    {
+        return [
+            'attributes from env var' => [
+                'foo=foo,bar=bar',
+                [],
+                ['foo' => 'foo', 'bar' => 'bar'],
+            ],
+            'user attributes have higher priority' => [
+                'foo=env-foo,bar=env-bar,baz=env-baz',
+                ['foo' => 'user-foo', 'bar' => 'user-bar'],
+                ['foo' => 'user-foo', 'bar' => 'user-bar', 'baz' => 'env-baz'],
+            ],
+        ];
+    }
+}
